# documentation/openApi.json
{
    "openapi": "3.0.0",
    "paths": {
...
      "/leitos/setores/filtro": {
        "get": {
          "operationId": "LeitosController_ocupacaoLeitosBySetorDescricao",
          "parameters": [
            {
              "name": "cdSetorAtendimento",
              "required": false,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "dsSetorAtendimento",
              "required": false,
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Seleciona informações de um leito de acordo com a Descrição informada",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LeitoDataGetDto"
                  }
                }
              }
            }
          },
          "tags": [
            "Leitos"
          ]
        }
      },
...
    },
    "info": {
      "title": "POC Tasy",
      "description": "Está API tem por objetivo, disponibilizar End-Points para envio dos dados do Censo Titularidade e Produto.",
      "version": "1.0",
      "contact": {
  
      }
    },
...
    "servers": [
      {
        "url": "https://tasy-poc.1lf2exw72bm8.us-south.codeengine.appdomain.cloud",
        "description": "URL Base da API"
      }
    ],
    "components": {
      "securitySchemes": {
        "Authorization": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
        }
      },
      "schemas": {
...
        "LeitoGetDto": {
          "type": "object",
          "properties": {
            "cdSetorAtendimento": {
              "type": "number"
            },
            "dsSetorAtendimento": {
              "type": "string"
            },
            "cdEstabelecimentoBase": {
              "type": "number"
            },
            "cdClassifSetor": {
              "type": "string"
            },
            "nmUnidade": {
              "type": "string"
            },
            "dsOcupHosp": {
              "type": "string"
            },
            "nrSeqApresentacao": {
              "type": "number"
            },
            "nrUnidadesSetor": {
              "type": "number"
            },
            "nrUnidadesTemporarias": {
              "type": "number"
            },
            "nrUnidadesOcupadas": {
              "type": "number"
            },
            "qtUnidadeAcomp": {
              "type": "number"
            },
            "nrUnidadesInterditadas": {
              "type": "number"
            },
            "nrUnidadesLivres": {
              "type": "number"
            },
            "nrUnidadesHigienizacao": {
              "type": "number"
            },
            "nrUnidadesReservadas": {
              "type": "number"
            },
            "qtUnidadesIsolamento": {
              "type": "number"
            },
            "qtUnidadesAlta": {
              "type": "number"
            },
            "qtUnidadeSaidaInter": {
              "type": "number"
            },
            "qtUnidadeChamadManut": {
              "type": "number"
            },
            "qtUnidadeManutencao": {
              "type": "number"
            },
            "nrUnidTempOcup": {
              "type": "number"
            },
            "nrUnidTempOcupadas": {
              "type": "number"
            },
            "qtUnidTempAcomp": {
              "type": "number"
            },
            "nrUnidTempInterditadas": {
              "type": "number"
            },
            "nrUnidTempLivres": {
              "type": "number"
            },
            "nrUnidTempHigienizacao": {
              "type": "number"
            },
            "nrUnidTempReservadas": {
              "type": "number"
            },
            "qtUnidTempIsolamento": {
              "type": "number"
            },
            "qtUnidTempSaidInterd": {
              "type": "number"
            },
            "qtUnidTempAlta": {
              "type": "number"
            },
            "nrUnidadesReserva": {
              "type": "number"
            },
            "ieOcupHospitalar": {
              "type": "string"
            },
            "ieSituacao": {
              "type": "string"
            },
            "qtDisponiveis": {
              "type": "number"
            },
            "qtPacIsolado": {
              "type": "number"
            },
            "porcentagem": {
              "type": "number"
            }
          },
          "required": [
            "cdSetorAtendimento",
            "porcentagem"
          ]
        },
        "LeitoDataGetDto": {
          "type": "object",
          "properties": {
            "data": {
              "$ref": "#/components/schemas/LeitoGetDto"
            }
          },
          "required": [
            "data"
          ]
        }
      }
    },
...
  }